SignerConfigurator
